chore: fix amplify-cli structure
