+++ package.json
@@ -60,6 +60,7 @@
     "eslint": "^6.8.0",
     "husky": "^3.0.3",
     "codecov": "^3.6.1",
+    "cpy-cli": "^3.0.0",
     "eslint-config-airbnb": "^18.0.1",
     "eslint-config-prettier": "^6.9.0",
     "eslint-plugin-import": "^2.20.0",
