+++ packages/amplify-cli/src/attach-backend-steps/a10-queryProvider.js
@@ -1,5 +1,5 @@
 const inquirer = require('inquirer');
-const { getProviderPlugins } = require('../../extensions/amplify-helpers/get-provider-plugins');
+const { getProviderPlugins } = require('../extensions/amplify-helpers/get-provider-plugins');
 const { normalizeProviderName } = require('../input-params-manager');
 
 async function run(context) {
