+++ packages/amplify-cli/src/attach-backend-steps/a40-generateFiles.js
@@ -1,10 +1,9 @@
 const fs = require('fs-extra');
 const sequential = require('promise-sequential');
-const { getFrontendPlugins } = require('../../extensions/amplify-helpers/get-frontend-plugins');
-const { getProviderPlugins } = require('../../extensions/amplify-helpers/get-provider-plugins');
-const gitManager = require('../../extensions/amplify-helpers/git-manager');
-const { initializeEnv } = require('../initialize-env');
-const { readJsonFile } = require('../../extensions/amplify-helpers/read-json-file');
+const { getFrontendPlugins } = require('../extensions/amplify-helpers/get-frontend-plugins');
+const { getProviderPlugins } = require('../extensions/amplify-helpers/get-provider-plugins');
+const gitManager = require('../extensions/amplify-helpers/git-manager');
+const { readJsonFile } = require('../extensions/amplify-helpers/read-json-file');
 
 async function run(context) {
   const { projectPath } = context.exeInfo.localEnvInfo;
@@ -19,15 +18,6 @@ async function run(context) {
   fs.ensureDirSync(dotConfigDirPath);
   fs.ensureDirSync(backendDirPath);
   fs.ensureDirSync(currentBackendDirPath);
-
-  // Get current-cloud-backend's amplify-meta
-  const currentAmplifyMetafilePath = amplify.pathManager.getCurrentAmplifyMetaFilePath();
-
-  let currentAmplifyMeta = {};
-
-  if (fs.existsSync(currentAmplifyMetafilePath)) {
-    currentAmplifyMeta = readJsonFile(currentAmplifyMetafilePath);
-  }
 
   const providerPlugins = getProviderPlugins(context);
   const providerOnSuccessTasks = [];
@@ -47,16 +37,18 @@ async function run(context) {
 
   await sequential(providerOnSuccessTasks);
 
-  await initializeEnv(context, currentAmplifyMeta);
-
-  if (!context.parameters.options.app) {
-    printWelcomeMessage(context);
+  const currentAmplifyMetafilePath = amplify.pathManager.getCurrentAmplifyMetaFilePath();
+  let currentAmplifyMeta = {};
+  if (fs.existsSync(currentAmplifyMetafilePath)) {
+    currentAmplifyMeta = readJsonFile(currentAmplifyMetafilePath);
   }
+  await context.amplify.onCategoryOutputsChange(context, currentAmplifyMeta);
+
+  return context;
 }
 
 function generateLocalRuntimeFiles(context) {
   generateLocalEnvInfoFile(context);
-  generateAmplifyMetaFile(context);
 }
 
 function generateLocalEnvInfoFile(context) {
@@ -64,17 +56,6 @@ function generateLocalEnvInfoFile(context) {
   const jsonString = JSON.stringify(context.exeInfo.localEnvInfo, null, 4);
   const localEnvFilePath = context.amplify.pathManager.getLocalEnvFilePath(projectPath);
   fs.writeFileSync(localEnvFilePath, jsonString, 'utf8');
-}
-
-function generateAmplifyMetaFile(context) {
-  if (context.exeInfo.isNewEnv) {
-    const { projectPath } = context.exeInfo.localEnvInfo;
-    const jsonString = JSON.stringify(context.exeInfo.amplifyMeta, null, 4);
-    const currentBackendMetaFilePath = context.amplify.pathManager.getCurrentAmplifyMetaFilePath(projectPath);
-    fs.writeFileSync(currentBackendMetaFilePath, jsonString, 'utf8');
-    const backendMetaFilePath = context.amplify.pathManager.getAmplifyMetaFilePath(projectPath);
-    fs.writeFileSync(backendMetaFilePath, jsonString, 'utf8');
-  }
 }
 
 function generateNonRuntimeFiles(context) {
@@ -85,7 +66,6 @@ function generateNonRuntimeFiles(context) {
 }
 
 function generateProjectConfigFile(context) {
-  // won't modify on new env
   if (context.exeInfo.isNewProject) {
     const { projectPath } = context.exeInfo.localEnvInfo;
     const jsonString = JSON.stringify(context.exeInfo.projectConfig, null, 4);
@@ -95,14 +75,19 @@ function generateProjectConfigFile(context) {
 }
 
 function generateProviderInfoFile(context) {
-  const { projectPath } = context.exeInfo.localEnvInfo;
-  let teamProviderInfo = {};
+  const { projectPath, envName } = context.exeInfo.localEnvInfo;
+  const { existingTeamProviderInfo, teamProviderInfo } = context.exeInfo;
   const providerInfoFilePath = context.amplify.pathManager.getProviderInfoFilePath(projectPath);
-  if (fs.existsSync(providerInfoFilePath)) {
-    teamProviderInfo = readJsonFile(providerInfoFilePath);
-    Object.assign(teamProviderInfo, context.exeInfo.teamProviderInfo);
-  } else {
-    ({ teamProviderInfo } = context.exeInfo);
+
+  if (existingTeamProviderInfo) {
+    if (existingTeamProviderInfo[envName]) {
+      if (existingTeamProviderInfo[envName].categories) {
+        teamProviderInfo[envName] = teamProviderInfo[envName] || {};
+        teamProviderInfo[envName].categories = existingTeamProviderInfo[envName].categories;
+      }
+      delete existingTeamProviderInfo[envName];
+    }
+    Object.assign(teamProviderInfo, existingTeamProviderInfo);
   }
 
   const jsonString = JSON.stringify(teamProviderInfo, null, 4);
@@ -110,9 +95,9 @@ function generateProviderInfoFile(context) {
 }
 
 function generateBackendConfigFile(context) {
-  if (context.exeInfo.isNewProject) {
-    const { projectPath } = context.exeInfo.localEnvInfo;
-    const backendConfigFilePath = context.amplify.pathManager.getBackendConfigFilePath(projectPath);
+  const { projectPath } = context.exeInfo.localEnvInfo;
+  const backendConfigFilePath = context.amplify.pathManager.getBackendConfigFilePath(projectPath);
+  if (!fs.existsSync(backendConfigFilePath)) {
     fs.writeFileSync(backendConfigFilePath, '{}', 'utf8');
   }
 }
@@ -125,24 +110,6 @@ function generateGitIgnoreFile(context) {
 
     gitManager.insertAmplifyIgnore(gitIgnoreFilePath);
   }
-}
-
-function printWelcomeMessage(context) {
-  context.print.info('');
-  context.print.success('Your project has been successfully initialized and connected to the cloud!');
-  context.print.info('');
-  context.print.success('Some next steps:');
-  context.print.info('"amplify status" will show you what you\'ve added already and if it\'s locally configured or deployed');
-  context.print.info('"amplify add <category>" will allow you to add features like user login or a backend API');
-  context.print.info('"amplify push" will build all your local backend resources and provision it in the cloud');
-  context.print.info('“amplify console” to open the Amplify Console and view your project status');
-  context.print.info(
-    '"amplify publish" will build all your local backend and frontend resources (if you have hosting category added) and provision it in the cloud'
-  );
-  context.print.info('');
-  context.print.success('Pro tip:');
-  context.print.info('Try "amplify add api" to create a backend API and then "amplify publish" to deploy everything');
-  context.print.info('');
 }
 
 module.exports = {
