+++ packages/amplify-cli/src/commands/init.js
@@ -1,11 +1,11 @@
-const preInitSetup = require('../lib/init-steps/preInitSetup');
-const postInitSetup = require('../lib/init-steps/postInitSetup');
-const analyzeProject = require('../lib/init-steps/s0-analyzeProject');
-const initFrontendHandler = require('../lib/init-steps/s1-initFrontend');
-const initProviders = require('../lib/init-steps/s2-initProviders');
-const onFailure = require('../lib/init-steps/s9-onFailure');
-const onSuccess = require('../lib/init-steps/s9-onSuccess');
-const { constructInputParams } = require('../lib/amplify-service-helper');
+const preInitSetup = require('../init-steps/preInitSetup');
+const postInitSetup = require('../init-steps/postInitSetup');
+const analyzeProject = require('../init-steps/s0-analyzeProject');
+const initFrontendHandler = require('../init-steps/s1-initFrontend');
+const initProviders = require('../init-steps/s2-initProviders');
+const onFailure = require('../init-steps/s9-onFailure');
+const onSuccess = require('../init-steps/s9-onSuccess');
+const { constructInputParams } = require('../amplify-service-helper');
 
 module.exports = {
   name: 'init',
