+++ packages/amplify-cli/src/config-steps/c1-configFrontend.js
@@ -1,5 +1,5 @@
 const inquirer = require('inquirer');
-const { getFrontendPlugins } = require('../../extensions/amplify-helpers/get-frontend-plugins');
+const { getFrontendPlugins } = require('../extensions/amplify-helpers/get-frontend-plugins');
 const { normalizeFrontendHandlerName } = require('../input-params-manager');
 
 async function run(context) {
