+++ packages/amplify-cli/src/config-steps/c9-onSuccess.js
