+++ packages/amplify-cli/src/extensions/amplify-helpers/push-resources.js
@@ -2,7 +2,7 @@ const fs = require('fs-extra');
 const { getProjectConfig } = require('./get-project-config');
 const { showResourceTable } = require('./resource-status');
 const { onCategoryOutputsChange } = require('./on-category-outputs-change');
-const { initializeEnv } = require('../../lib/initialize-env');
+const { initializeEnv } = require('../../initialize-env');
 const { getProviderPlugins } = require('./get-provider-plugins');
 const { getEnvInfo } = require('./get-env-info');
 const { readJsonFile } = require('./read-json-file');
