+++ packages/amplify-cli/src/global-prefix.js
