+++ packages/amplify-cli/src/init-steps/postInitSetup.js
