+++ packages/amplify-cli/src/init-steps/s2-initProviders.js
@@ -1,6 +1,6 @@
 const inquirer = require('inquirer');
 const sequential = require('promise-sequential');
-const { getProviderPlugins } = require('../../extensions/amplify-helpers/get-provider-plugins');
+const { getProviderPlugins } = require('../extensions/amplify-helpers/get-provider-plugins');
 const { normalizeProviderName } = require('../input-params-manager');
 
 async function run(context) {
