+++ packages/amplify-cli/templates/amplify-skeleton/.config/project-config.json
