+++ packages/amplify-cli/templates/amplify-skeleton/backend/api/amplifyDatasource/schema.graphql
