+++ packages/amplify-cli/templates/amplify-skeleton/backend/api/amplifyDatasource/stacks/CustomResources.json
