+++ packages/amplify-cli/templates/amplify-skeleton/backend/api/amplifyDatasource/transform.conf.json
