+++ packages/amplify-cli/templates/amplify-skeleton/backend/backend-config.json
