+++ tsconfig.json
@@ -63,6 +63,7 @@
 
   },
   "exclude": [
+    "**/__mocks__/**/*",
     "**/__tests__/**/*",
     "**/lib/**/*"
   ]
